@@ -1,25 +1,18 @@
 source 'https://rubygems.org'
+ruby '2.4.1'
 
-ruby "2.1.0"
-#ruby-gemset=csvlint
-
-gem 'rails', '~> 4.0.12'
+gem 'rake', '~> 11.0'
+gem 'rails', '~> 4.2.8'
 gem 'dotenv-rails'
 
-# Sprockets pinned to avoid a problem in 2.12.3 with our stylesheets
-gem "sprockets", "~> 2.11.3"
-
-# Use sqlite3 as the database for Active Record
-gem 'sqlite3'
-
 # Use SCSS for stylesheets
-gem 'sass-rails', '~> 4.0.0'
+gem 'sass-rails', '~> 5.0'
 
 # Use Uglifier as compressor for JavaScript assets
 gem 'uglifier', '>= 1.3.0'
 
 # Use CoffeeScript for .js.coffee assets and views
-gem 'coffee-rails', '~> 4.0.0'
+gem 'coffee-rails', '~> 4.1.0'
 
 # See https://github.com/sstephenson/execjs#readme for more supported runtimes
 gem 'therubyracer', platforms: :ruby
@@ -31,21 +24,16 @@ gem 'jquery-rails'
 gem 'turbolinks'
 
 # Build JSON APIs with ease. Read more: https://github.com/rails/jbuilder
-gem 'jbuilder' #, '~> 1.2'
+gem 'jbuilder', '~> 2.0'
 
-gem 'bootstrap-sass', '~> 3.1.1'
+gem 'sdoc', '~> 0.4.0', group: :doc
 
-group :doc do
-  # bundle exec rake doc:rails generates the API under doc/api.
-  gem 'sdoc', require: false
-end
-
-
-gem 'foreman', "< 0.65.0"
+gem 'foreman'
 
 group :production do
   gem 'thin'
-  gem 'mysql2'
+  gem 'rails_12factor'
+  gem 'puma'
 end
 
 group :development, :test do
@@ -65,10 +53,14 @@ group :development, :test do
   gem 'poltergeist'
   gem 'vcr'
   gem 'cucumber-api-steps', require: false, github: 'theodi/cucumber-api-steps', branch: 'feature-test-content-type'
+  # Use sqlite3 as the database for Active Record
+  gem 'sqlite3'
 end
 
 group :development do
   gem 'travis'
+  gem 'web-console', '~> 2.0'
+  gem 'spring'
 end
 
 # Use ActiveModel has_secure_password
@@ -80,27 +72,30 @@ end
 # Use Capistrano for deployment
 # gem 'capistrano', group: :development
 
-# Use debugger
-# gem 'debugger', group: [:development, :test]
-
+gem 'bootstrap-sass', '~> 3.1.1'
 gem 'rack-google-analytics'
-gem 'mongoid'
-gem 'mongoid-grid_fs', github: 'ahoward/mongoid-grid_fs'
+gem 'mongoid', '~> 5.0'
+gem 'bson', '3.1.1'
+gem 'mongoid-grid_fs', '~> 2.2'
 gem 'bson_ext'
 gem 'kaminari'
 gem 'bootstrap-kaminari-views'
-gem 'data_kitten', github: 'theodi/data_kitten', ref: "e343510bd15e3329c1f2fab35035e248195348be"
+gem 'data_kitten', github: 'theodi/data_kitten', ref: "e343510bd15e3329c1f2fab35035e248195348be", require: false
 gem 'rubyzip'
 
-gem 'csvlint', github: 'theodi/csvlint.rb'
+gem 'csvlint'
 gem 'rest-client'
 
 gem 'nokogiri', '~> 1.5'
 
 gem 'airbrake'
 gem 'font-awesome-rails'
-gem 'delayed_job_mongoid', github: 'collectiveidea/delayed_job_mongoid'
+gem 'sidekiq', '= 4.2.4'
 gem 'data_uri'
 gem 'jquery-dotdotdot-rails'
-gem 'alternate_rails', github: 'theodi/alternate-rails'
+gem 'alternate_rails', github: 'theodi/alternate-rails', ref: 'v4.2.0'
 gem 'rack-cors'
+
+gem 'resumable_upload', github: "theodi/resumable-upload"
+gem 'pusher'
+gem 'cloudflare'
