@@ -1,41 +1,47 @@
 require 'spec_helper'
 
 describe PackageController, type: :controller do
-  
+
+  before(:each) do
+    mock_client = double(Pusher::Channel)
+    allow(Pusher).to receive(:[]) { mock_client }
+    allow(mock_client).to receive(:trigger)
+  end
+
   describe "POST 'create'" do
 
     it "redirects to root if no URL is supplied" do
       post 'create'
-      response.should be_redirect
-      response.location.should == root_url
+      expect(response).to be_redirect
+      expect(response.location).to eq(root_url)
     end
 
     it "redirects to root if an invalid url is supplied" do
       post 'create', urls: ['complete@balls:/']
-      response.should be_redirect
-      response.location.should == root_url
+      expect(response).to be_redirect
+      expect(response.location).to eq(root_url)
     end
 
     it "redirects to root if a relative url is supplied" do
       post 'create', urls: ['/test.csv']
-      response.should be_redirect
-      response.location.should == root_url
+      expect(response).to be_redirect
+      expect(response.location).to eq(root_url)
     end
 
     it "redirects to root if a non-http[s] url is supplied" do
       post 'create', urls: ['file://test.csv']
-      response.should be_redirect
-      response.location.should == root_url
+      expect(response).to be_redirect
+      expect(response.location).to eq(root_url)
     end
 
     it "creates a validation and redirects if a sensible url is supplied" do
       mock_file("http://example.com/test.csv", 'csvs/valid.csv')
       post 'create', urls: ['http://example.com/test.csv']
-      response.should be_redirect
+      expect(response).to be_redirect
       validation = Validation.first
-      response.location.should == validation_url(validation)
+      expect(response.location).to eq(validation_url(validation))
     end
-    
+
     it "supports multiple urls" do
        mock_file("http://example.com/test.csv", 'csvs/valid.csv')
        mock_file("http://example.com/test2.csv", 'csvs/valid.csv')
@@ -45,24 +51,24 @@ describe PackageController, type: :controller do
                               'http://example.com/test2.csv',
                               'http://example.com/test3.csv'
                             ]
-       response.should be_redirect
+       expect(response).to be_redirect
        package = Package.first
-       package.validations.count.should == 3
-       response.location.should == package_url(package)
+       expect(package.validations.count).to eq(3)
+       expect(response.location).to eq(package_url(package))
     end
-    
+
     it "supports multiple files" do
-      post 'create', files: [
-                        mock_upload('csvs/valid.csv'),
-                        mock_upload('csvs/valid.csv'),
-                        mock_upload('csvs/valid.csv')
-                      ]                
-      response.should be_redirect
+      post 'create', file_ids: [
+                        mock_upload('valid.csv'),
+                        mock_upload('valid.csv'),
+                        mock_upload('valid.csv')
+                      ]
+      expect(response).to be_redirect
       package = Package.first
-      package.validations.count.should == 3      
-      response.location.should == package_url(package)
+      expect(package.validations.count).to eq(3)
+      expect(response.location).to eq(package_url(package))
     end
-    
+
     it "supports multiple zip urls" do
        mock_file("http://example.com/valid.zip", 'csvs/valid.zip')
        mock_file("http://example.com/multiple_files.zip", 'csvs/multiple_files.zip')
@@ -70,34 +76,34 @@ describe PackageController, type: :controller do
                               'http://example.com/valid.zip',
                               'http://example.com/multiple_files.zip',
                             ]
-       response.should be_redirect
+       expect(response).to be_redirect
        package = Package.first
-       package.validations.count.should == 4
-       response.location.should == package_url(package)
+       expect(package.validations.count).to eq(4)
+       expect(response.location).to eq(package_url(package))
     end
-    
+
     it "supports multiple zip files" do
-      post 'create', files: [
-                        mock_upload('csvs/valid.zip', 'application/zip'),
-                        mock_upload('csvs/multiple_files.zip', 'application/zip'),
+      post 'create', file_ids: [
+                        mock_upload('valid.zip'),
+                        mock_upload('multiple_files.zip'),
                       ]
-      response.should be_redirect
+      expect(response).to be_redirect
       package = Package.first
-      package.validations.count.should == 4
-      response.location.should == package_url(package)
+      expect(package.validations.count).to eq(4)
+      expect(response.location).to eq(package_url(package))
     end
-    
+
     it "supports data URLs" do
       post 'create', files_data: [
                             create_data_uri('csvs/valid.csv')
                           ]
-            
-      response.should be_redirect
+
+      expect(response).to be_redirect
       package = Package.first
-      package.validations.count.should == 1
-      response.location.should == validation_url(package.validations.first)
+      expect(package.validations.count).to eq(1)
+      expect(response.location).to eq(validation_url(package.validations.first))
     end
-    
+
     it "supports multiple data URLs" do
       post 'create', files_data: [
                             create_data_uri('csvs/valid.csv'),
@@ -105,34 +111,34 @@ describe PackageController, type: :controller do
                             create_data_uri('csvs/valid.csv'),
                             create_data_uri('csvs/valid.csv')
                           ]
-            
-      response.should be_redirect
+
+      expect(response).to be_redirect
       package = Package.first
-      package.validations.count.should == 4
-      response.location.should == package_url(package)
+      expect(package.validations.count).to eq(4)
+      expect(response.location).to eq(package_url(package))
     end
-    
+
     it "supports single zip files as data URLs" do
       post 'create', files_data: [
                         create_data_uri('csvs/valid.zip', 'application/zip'),
                       ]
-      response.should be_redirect
+      expect(response).to be_redirect
       package = Package.first
-      package.validations.count.should == 1
-      response.location.should == validation_url(package.validations.first)
+      expect(package.validations.count).to eq(1)
+      expect(response.location).to eq(validation_url(package.validations.first))
     end
-    
+
     it "supports multiple zip files as data URLs" do
       post 'create', files_data: [
                         create_data_uri('csvs/valid.zip', 'application/zip'),
                         create_data_uri('csvs/multiple_files.zip', 'application/zip'),
                       ]
-      response.should be_redirect
+      expect(response).to be_redirect
       package = Package.first
-      package.validations.count.should == 4
-      response.location.should == package_url(package)
+      expect(package.validations.count).to eq(4)
+      expect(response.location).to eq(package_url(package))
     end
-    
+
     it "supports schema uploads as data URLs" do
       mock_file("http://example.com/test.csv", 'csvs/all_constraints.csv')
       post 'create', urls: [
@@ -140,55 +146,57 @@ describe PackageController, type: :controller do
                            ],
                      schema: "1",
                      schema_data: create_data_uri('schemas/all_constraints.json', 'application/json')
-                     
-      response.should be_redirect
+      # above accurately emulates how the file upload works by a user
+      expect(response).to be_redirect
       package = Package.first
       validation = package.validations.first
       validator = validation.validator
-      response.location.should == validation_url(validation)
-      validator.errors.count.should == 10
-      validator.errors[0].type.should == :missing_value
-      validator.errors[1].type.should == :min_length
-      validator.errors[2].type.should == :min_length
-      validator.errors[3].type.should == :max_length
-      validator.errors[4].type.should == :pattern
-      validator.errors[5].type.should == :unique
-      validator.errors[6].type.should == :below_minimum
-      validator.errors[7].type.should == :above_maximum
-      validator.errors[8].type.should == :above_maximum
-      validator.errors[9].type.should == :below_minimum
+      expect(response.location).to eq(validation_url(validation))
+      # byebug
+      expect(validator.errors.count).to eq(10)
+      # the above is breaking at present
+      expect(validator.errors[0].type).to eq(:missing_value)
+      expect(validator.errors[1].type).to eq(:min_length)
+      expect(validator.errors[2].type).to eq(:min_length)
+      expect(validator.errors[3].type).to eq(:max_length)
+      expect(validator.errors[4].type).to eq(:pattern)
+      expect(validator.errors[5].type).to eq(:unique)
+      expect(validator.errors[6].type).to eq(:below_minimum)
+      expect(validator.errors[7].type).to eq(:above_maximum)
+      expect(validator.errors[8].type).to eq(:above_maximum)
+      expect(validator.errors[9].type).to eq(:below_minimum)
     end
 
   end
 
   describe "POST 'create' HTML" do
-    
+
     it "has no warnings or errors for valid CSV" do
       mock_file("http://example.com/test.csv", 'csvs/valid.csv')
       post 'create', urls: ['http://example.com/test.csv']
-      response.should be_redirect
+      expect(response).to be_redirect
       validation = Marshal.load(Validation.first.result)
-      validation.warnings.should be_empty
-      validation.errors.should be_empty
+      expect(validation.warnings).to be_empty
+      expect(validation.errors).to be_empty
     end
-    
+
     it "has warnings or errors for warning CSV" do
       mock_file("http://example.com/test.csv", 'csvs/warnings.csv')
       post 'create', urls: ['http://example.com/test.csv']
-      response.should be_redirect
+      expect(response).to be_redirect
       validation = Marshal.load(Validation.first.result)
-      validation.warnings.should_not be_empty
-      validation.errors.should be_empty
+      expect(validation.warnings).not_to be_empty
+      expect(validation.errors).to be_empty
     end
-    
+
     it "has errors for error CSV" do
       mock_file("http://example.com/test.csv", 'csvs/errors.csv')
       post 'create', urls: ['http://example.com/test.csv']
-      response.should be_redirect      
+      expect(response).to be_redirect
       validation = Marshal.load(Validation.first.result)
-      validation.errors.should_not be_empty
+      expect(validation.errors).not_to be_empty
     end
-    
+
   end
 
 end
