@@ -2,13 +2,46 @@ require 'spec_helper'
 
 describe ValidationController, type: :controller do
 
+  describe "GET 'index'" do
+
+    it 'generates a CSV of all validations with a url' do
+      [
+        "valid",
+        "warnings",
+        "invalid",
+        "not_found",
+      ].each_with_index do |state, i|
+        (i + 1).times { |i| FactoryGirl.create :validation, url: "http://data.com/data#{i}.csv", state: state }
+      end
+
+      5.times { FactoryGirl.create :validation, url: nil }
+
+      get 'index', format: "csv"
+
+      expect(response.content_type).to eq('text/csv; charset=utf-8; header=present')
+
+      csv = CSV.parse(response.body)
+      expect(csv.count).to eq(11)
+    end
+
+  end
+
   describe "POST 'update'" do
+
+    before(:each) do
+      @connection = double(CloudFlare::Connection)
+
+      allow(CloudFlare::Connection).to receive(:new) {
+        allow(@connection).to receive(:zone_file_purge)
+        @connection
+      }
+    end
 
     it "updates a CSV sucessfully" do
        mock_file("http://example.com/test.csv", 'csvs/valid.csv')
        Validation.create_validation('http://example.com/test.csv')
        put 'update', id: Validation.first.id
-       response.should be_redirect
+       expect(response).to be_redirect
     end
 
     it "updates a CSV with a new schema sucessfully" do
@@ -27,9 +60,16 @@ describe ValidationController, type: :controller do
        put 'update', params
 
        validator = Marshal.load Validation.first.result
-       validator.warnings.select { |warning| warning.type == :check_options }.count.should == 0
+       expect(validator.warnings.select { |warning| warning.type == :check_options }.count).to eq(0)
 
-       response.should be_redirect
+       expect(response).to be_redirect
+    end
+
+    it 'purges the cache when updating' do
+      mock_file("http://example.com/test.csv", 'csvs/valid.csv')
+      Validation.create_validation('http://example.com/test.csv')
+      expect(@connection).to receive(:zone_file_purge).with('csvlint.io', validation_url(Validation.first))
+      put 'update', id: Validation.first.id
     end
 
   end
@@ -41,8 +81,8 @@ describe ValidationController, type: :controller do
       Validation.create_validation('http://example.com/test.csv')
       validation = Validation.first
       get 'show', id: validation.id, format: :png
-      response.should be_success
-      response.body.length.should == 1588
+      expect(response).to be_success
+      expect(response.body.length).to eq(1588)
     end
 
     it "returns invalid image for a CSV with errors" do
@@ -50,8 +90,8 @@ describe ValidationController, type: :controller do
       Validation.create_validation('http://example.com/test.csv')
       validation = Validation.first
       get 'show', id: validation.id, format: :png
-      response.should be_success
-      response.body.length.should == 1760
+      expect(response).to be_success
+      expect(response.body.length).to eq(1760)
     end
 
     it "returns warning image for a CSV with warnings" do
@@ -59,8 +99,8 @@ describe ValidationController, type: :controller do
       Validation.create_validation('http://example.com/test.csv')
       validation = Validation.first
       get 'show', id: validation.id, format: :png
-      response.should be_success
-      response.body.length.should == 2099
+      expect(response).to be_success
+      expect(response.body.length).to eq(2099)
     end
 
   end
@@ -72,8 +112,8 @@ describe ValidationController, type: :controller do
       Validation.create_validation('http://example.com/test.csv')
       validation = Validation.first
       get 'show', id: validation.id, format: :svg
-      response.should be_success
-      response.body.should include(">valid<")
+      expect(response).to be_success
+      expect(response.body).to include(">valid<")
     end
 
     it "returns invalid image for a CSV with errors" do
@@ -81,8 +121,8 @@ describe ValidationController, type: :controller do
       Validation.create_validation('http://example.com/test.csv')
       validation = Validation.first
       get 'show', id: validation.id, format: :svg
-      response.should be_success
-      response.body.should include(">invalid<")
+      expect(response).to be_success
+      expect(response.body).to include(">invalid<")
     end
 
     it "returns warning image for a CSV with warnings" do
@@ -90,26 +130,27 @@ describe ValidationController, type: :controller do
       Validation.create_validation('http://example.com/test.csv')
       validation = Validation.first
       get 'show', id: validation.id, format: :svg
-      response.should be_success
-      response.body.should include(">warnings<")
+      expect(response).to be_success
+      expect(response.body).to include(">warnings<")
     end
 
     it "queues another check when the image is loaded" do
       mock_file("http://example.com/test.csv", 'csvs/valid.csv')
       Validation.create_validation('http://example.com/test.csv')
       validation = Validation.first
-      Validation.any_instance.should_receive(:delay).and_call_original
-      get 'show', id: validation.id, format: :svg
-      Delayed::Job.count.should == 1
+      expect {
+        get 'show', id: validation.id, format: :svg
+      }.to change(Sidekiq::Extensions::DelayedClass.jobs, :size).by(1)
     end
 
     it "doesn't queue another check when the image is loaded if revalidate is false" do
       mock_file("http://example.com/test.csv", 'csvs/valid.csv')
       Validation.create_validation('http://example.com/test.csv')
       validation = Validation.first
-      Validation.any_instance.should_not_receive(:delay)
-      get 'show', id: validation.id, format: :svg, revalidate: false
-      Delayed::Job.count.should == 0
+      expect(Validation).to_not receive(:delay)
+      expect {
+        get 'show', id: validation.id, format: :svg, revalidate: false
+      }.to change(Sidekiq::Extensions::DelayedClass.jobs, :size).by(0)
     end
 
   end
