@@ -1,12 +1,38 @@
+require 'package_processor'
+require 'lite_kitten/lite_kitten'
+
 class LocalDataset < DataKitten::Dataset
   extend DataKitten::PublishingFormats::Datapackage
 
   def origin
     :local
   end
-  
+
   def publishing_format
+    # A 'package' can be a DataPackage, or it can be any collection of multiple CSVs,
+    # and/or collection of CSVs and their schema metadata
+    # ?? is this used in conjunction with package_helper
     :datapackage
+  end
+end
+
+class RemoteDataset < DataKitten::Dataset
+
+  def initialize(url)
+    @access_url = url
+    detect_publishing_format
+  end
+
+  def detect_publishing_format
+     [
+       DataKitten::PublishingFormats::Datapackage,
+       DataKitten::PublishingFormats::CKAN
+     ].each do |format|
+       if format.supported?(self)
+         extend format
+         break
+       end
+    end
   end
 end
 
@@ -31,54 +57,48 @@ class Package
     return attributes
   end
 
+  def self.process(id, params)
+    PackageProcessor.new(params, id).process
+  end
+
   def create_package(sources, schema_url = nil, schema = nil)
-    return nil if sources.count == 0    
-    
-    if sources.first.class == Hash
-      sources.map! { |io| 
-        {
-          :body => Mongoid::GridFs.get(io[:csv_id]).data,
-          :csv_id => io[:csv_id],
-          :filename => io[:filename]
-        }  
-      }
-    end
-            
+    return nil if sources.count == 0
+
     if sources.count == 1 && possible_package?(sources.first)
       dataset = create_dataset(sources.first)
       return create_datapackage(dataset) unless dataset.nil?
-    end    
+    end
 
     update_attributes({ type: set_type(sources) })
 
     sources.each do |source|
       validations << Validation.create_validation(source, schema_url, schema)
     end
-  
+
     save
     self
   end
-  
+
   def create_dataset(source)
     if source.respond_to?(:body)
       dataset = LocalDataset.new(access_url: source.string_io)
     else
-      dataset = DataKitten::Dataset.new(access_url: source)
+      dataset = RemoteDataset.new(source)
     end
     return nil unless [:ckan, :datapackage].include? dataset.publishing_format
     dataset
   end
-  
-  def create_datapackage(dataset)  
+
+  def create_datapackage(dataset)
     validations = create_validations(dataset)
-    
+
     return nil if validations.count == 0
-    
+
     update_attributes( parse_package(dataset, validations) )
     save
     self
   end
-  
+
   def create_validations(dataset)
     validations = []
     dataset.distributions.each do |distribution|
@@ -88,17 +108,17 @@ class Package
     end
    validations
   end
-  
+
   def possible_package?(source)
     source.class == String || local_package?( source )
   end
-  
+
   def local_package?(source)
     source.respond_to?(:string_io) && source.filename =~ /datapackage\.json/
-  end  
-  
+  end
+
   def set_type(sources)
-    return "files" if sources.first.respond_to?(:tempfile) 
+    return "files" if sources.first.respond_to?(:tempfile)
     return "urls" if sources.first.class == String
   end
 
@@ -109,7 +129,7 @@ class Package
 
   def create_schema(distribution)
     unless distribution.schema.nil?
-      schema = Csvlint::Schema.from_json_table(nil, distribution.schema) 
+      schema = Csvlint::Schema.from_json_table(nil, distribution.schema)
     end
     return schema
   end
