@@ -1,5 +1,7 @@
+require 'package_processor'
+
 Then(/^my CSV should be placed in a background job$/) do
-  Package.any_instance.should_receive(:delay).and_call_original
+  Package.should_receive(:delay).and_call_original
 end
 
 When(/^I wait for the ajax request to finish$/) do
@@ -10,7 +12,7 @@ When(/^I wait for the ajax request to finish$/) do
 end
 
 When(/^the CSV has finished processing$/) do
-  Delayed::Job.first.invoke_job
+  Sidekiq::Extensions::DelayedClass.drain
   sleep 5
 end
 
@@ -18,6 +20,12 @@ When(/^I wait for the package to be created$/) do
   patiently do
     Package.first.should_not be_nil
   end
+end
+
+Then(/^a Pusher notification should be sent$/) do
+  mock_client = double(Pusher::Channel)
+  expect(Pusher).to receive(:[]) { mock_client }
+  expect(mock_client).to receive(:trigger)
 end
 
 Then(/^I should be redirected to my validation results$/) do
